% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlpiper.R
\name{process_async}
\alias{process_async}
\title{Process a text with NLPipe}
\usage{
process_async(module, texts, server = getOption("nlpiper.server", default =
  "http://localhost:5001"), ids = NULL, as_ascii = F)
}
\arguments{
\item{module}{Name of the NLPipe module to call (e.g. test_upper, corenlp_lemmatize)}

\item{texts}{Texts to process}

\item{server}{NLPipe server or local folder (default: localhost:5001)}

\item{ids}{optional explicit ids}

\item{as_ascii}{if TRUE, force convert text to ascii before processing}
}
\value{
IDs of the tasks
}
\description{
For instructions for setting up an NLPipe server, see \url{https://github.com/vanatteveldt/nlpipe}
}
\examples{
\dontrun{
## setup corenlp, e.g. run corenlp docker.
## For instructions see  <https://github.com/vanatteveldt/nlpipe>

## start process
task_id = nlpiper::process_async("corenlp_lemmatize", c("This is an example.", "This as well."))

## check status
nlpiper::status("corenlp_lemmatize", ids = task_id)

## retrieve results
nlpiper::result('corenlp_lemmatize', ids = task_id, format = 'csv')
}
}
