% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlpiper.R
\name{result}
\alias{result}
\title{Fetch NLPipe results}
\usage{
result(module, ids, server = getOption("nlpiper.server", default =
  "http://localhost:5001"), format = NULL)
}
\arguments{
\item{module}{Name of the NLPipe module to call (e.g. test_upper, corenlp_lemmatize)}

\item{ids}{IDs of the task to get results for}

\item{server}{NLPipe server or local folder (default: localhost:5001)}

\item{format}{The format to download results as (e.g. csv)}
}
\value{
The processed text(s), or NA if status was not 'DONE'. If format is csv, return a single data frame with all result lines
}
\description{
Fetch NLPipe results
}
\examples{
\dontrun{
## setup corenlp, e.g. run corenlp docker.
## For instructions see  <https://github.com/vanatteveldt/nlpipe>

## start process
task_id = nlpiper::process_async("corenlp_lemmatize", c("This is an example.", "This as well."))

## check status
nlpiper::status("corenlp_lemmatize", ids = task_id)

## retrieve results
nlpiper::result('corenlp_lemmatize', ids = task_id, format = 'csv')
}
}
