% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlpiper.R
\name{status}
\alias{status}
\title{Check NLPipe processing status}
\usage{
status(module, ids, server = getOption("nlpiper.server", default =
  "http://localhost:5001"))
}
\arguments{
\item{module}{Name of the NLPipe module to call (e.g. test_upper, corenlp_lemmatize)}

\item{ids}{IDs of the task to check status}

\item{server}{NLPipe server or local folder (default: localhost:5001)}
}
\value{
A string for each task indicating processing status
}
\description{
Check NLPipe processing status
}
\examples{
\dontrun{
## setup corenlp, e.g. run corenlp docker.
## For instructions see  <https://github.com/vanatteveldt/nlpipe>

## start process
task_id = nlpiper::process_async("corenlp_lemmatize", c("This is an example.", "This as well."))

## check status
nlpiper::status("corenlp_lemmatize", ids = task_id)

## retrieve results
nlpiper::result('corenlp_lemmatize', ids = task_id, format = 'csv')
}
}
