% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlpiper.R
\name{process}
\alias{process}
\title{Process a text with NLPipe and wait for result}
\usage{
process(module, text, server = getOption("nlpiper.server", default =
  "http://localhost:5001"), format = NULL, ids = NULL, as_ascii = F)
}
\arguments{
\item{module}{Name of the NLPipe module to call (e.g. test_upper, corenlp_lemmatize)}

\item{text}{Text to process}

\item{server}{NLPipe server or local folder (default: localhost:5001)}

\item{format}{The format to download results as (e.g. csv)}

\item{ids}{optinal explicit ids}

\item{as_ascii}{if TRUE, force convert text to ascii before processing}
}
\value{
The processed text
}
\description{
Process a text with NLPipe and wait for result
}
\examples{
\dontrun{
## setup corenlp, e.g. run corenlp docker.
## For instructions see  https://github.com/vanatteveldt/nlpipe

nlpiper::process("test_upper", "test")
nlpiper::process("corenlp_lemmatize", "This is an example.", format = 'csv')
}
}
