% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlpiper.R
\name{process}
\alias{process}
\title{Process a text with NLPipe and wait for result}
\usage{
process(module, text, server = getOption("nlpiper.server", default =
  "http://localhost:5001"), format = NULL)
}
\arguments{
\item{module}{Name of the NLPipe module to call (e.g. test_upper, corenlp_lemmatize)}

\item{text}{Text to process}

\item{server}{NLPipe server or local folder (default: localhost:5001)}

\item{format}{The format to download results as (e.g. csv)}
}
\value{
The processed text
}
\description{
Process a text with NLPipe and wait for result
}

