% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlpiper.R
\name{process_async}
\alias{process_async}
\title{Process a text with NLPipe}
\usage{
process_async(module, text, server = getOption("nlpiper.server", default =
  "http://localhost:5001"))
}
\arguments{
\item{module}{Name of the NLPipe module to call (e.g. test_upper, corenlp_lemmatize)}

\item{text}{Text to process}

\item{server}{NLPipe server or local folder (default: localhost:5001)}
}
\value{
The ID of the task
}
\description{
Process a text with NLPipe
}

