% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlpiper.R
\name{result}
\alias{result}
\title{Fetch NLPipe results}
\usage{
result(module, id, server = getOption("nlpiper.server", default =
  "http://localhost:5001"))
}
\arguments{
\item{module}{Name of the NLPipe module to call (e.g. test_upper, corenlp_lemmatize)}

\item{id}{ID of the task to get results for}

\item{server}{NLPipe server or local folder (default: localhost:5001)}
}
\value{
The processed text
}
\description{
Fetch NLPipe results
}

