% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlpiper.R
\name{status}
\alias{status}
\title{Check NLPipe processing status}
\usage{
status(module, id, server = getOption("nlpiper.server", default =
  "http://localhost:5001"))
}
\arguments{
\item{module}{Name of the NLPipe module to call (e.g. test_upper, corenlp_lemmatize)}

\item{id}{ID of the task to check status}

\item{server}{NLPipe server or local folder (default: localhost:5001)}
}
\value{
A string indicating processing status
}
\description{
Check NLPipe processing status
}

